test: split ACP manager failover coverage
