+++ src/acp/control-plane/manager.test-helpers.ts
@@ -1,0 +1,356 @@
+import type { AcpRuntime, AcpRuntimeCapabilities } from "@openclaw/acp-core/runtime/types";
+import { afterEach, beforeEach, expect, vi } from "vitest";
+import type { OpenClawConfig } from "../../config/config.js";
+import type { AcpSessionRuntimeOptions, SessionAcpMeta } from "../../config/sessions/types.js";
+import { resetHeartbeatWakeStateForTests } from "../../infra/heartbeat-wake.js";
+import { withTempDir } from "../../test-helpers/temp-dir.js";
+import { resetAcpActiveTurnsForTests } from "./active-turns.js";
+
+export type { AcpRuntime, OpenClawConfig, SessionAcpMeta };
+
+const hoistedMocks = vi.hoisted(() => {
+  const listAcpSessionEntriesMock = vi.fn();
+  const readAcpSessionEntryMock = vi.fn();
+  const upsertAcpSessionMetaMock = vi.fn();
+  const getAcpRuntimeBackendMock = vi.fn();
+  const requireAcpRuntimeBackendMock = vi.fn();
+  return {
+    listAcpSessionEntriesMock,
+    readAcpSessionEntryMock,
+    upsertAcpSessionMetaMock,
+    getAcpRuntimeBackendMock,
+    requireAcpRuntimeBackendMock,
+  };
+});
+
+vi.mock("../runtime/session-meta.js", () => ({
+  listAcpSessionEntries: (params: unknown) => hoistedMocks.listAcpSessionEntriesMock(params),
+  readAcpSessionEntry: (params: unknown) => hoistedMocks.readAcpSessionEntryMock(params),
+  upsertAcpSessionMeta: (params: unknown) => hoistedMocks.upsertAcpSessionMetaMock(params),
+}));
+
+vi.mock("../runtime/registry.js", () => ({
+  getAcpRuntimeBackend: (backendId?: string) => hoistedMocks.getAcpRuntimeBackendMock(backendId),
+  requireAcpRuntimeBackend: (backendId?: string) =>
+    hoistedMocks.requireAcpRuntimeBackendMock(backendId),
+}));
+
+export const hoisted = hoistedMocks;
+
+export const {
+  AcpSessionManager,
+  testing: { resetAcpSessionManagerForTests },
+} = await import("./manager.js");
+export const { AcpRuntimeError } = await import("../runtime/errors.js");
+export const { findTaskByRunId, resetTaskRegistryForTests } =
+  await import("../../tasks/task-registry.js");
+export const { resetTaskFlowRegistryForTests } = await import("../../tasks/task-flow-registry.js");
+export const { configureTaskFlowRegistryRuntime } =
+  await import("../../tasks/task-flow-registry.store.js");
+export const { installInMemoryTaskRegistryRuntime } =
+  await import("../../test-utils/task-registry-runtime.js");
+
+export const baseCfg = {
+  acp: {
+    enabled: true,
+    backend: "acpx",
+    dispatch: { enabled: true },
+  },
+} as const;
+export const ORIGINAL_STATE_DIR = process.env.OPENCLAW_STATE_DIR;
+
+export async function withAcpManagerTaskStateDir(
+  run: (root: string) => Promise<void>,
+): Promise<void> {
+  await withTempDir({ prefix: "openclaw-acp-manager-task-" }, async (root) => {
+    process.env.OPENCLAW_STATE_DIR = root;
+    resetTaskRegistryForTests({ persist: false });
+    resetTaskFlowRegistryForTests({ persist: false });
+    installInMemoryTaskRegistryRuntime();
+    configureTaskFlowRegistryRuntime({
+      store: {
+        loadSnapshot: () => ({
+          flows: new Map(),
+        }),
+        saveSnapshot: () => {},
+        upsertFlow: () => {},
+        deleteFlow: () => {},
+        close: () => {},
+      },
+    });
+    try {
+      await run(root);
+    } finally {
+      resetTaskRegistryForTests({ persist: false });
+      resetTaskFlowRegistryForTests({ persist: false });
+    }
+  });
+}
+
+export async function flushMicrotasks(rounds = 3): Promise<void> {
+  for (let index = 0; index < rounds; index += 1) {
+    await Promise.resolve();
+  }
+}
+
+export function createDeferred(): { promise: Promise<void>; resolve: () => void } {
+  let resolve: (() => void) | undefined;
+  const promise = new Promise<void>((next) => {
+    resolve = next;
+  });
+  if (!resolve) {
+    throw new Error("Expected deferred resolver to be initialized");
+  }
+  return { promise, resolve };
+}
+
+export function expectRecordFields(record: unknown, expected: Record<string, unknown>) {
+  if (!record || typeof record !== "object") {
+    throw new Error("Expected record");
+  }
+  const actual = record as Record<string, unknown>;
+  for (const [key, value] of Object.entries(expected)) {
+    expect(actual[key]).toEqual(value);
+  }
+  return actual;
+}
+
+export async function expectRejectedRecord(
+  promise: Promise<unknown>,
+  expected: Record<string, unknown>,
+) {
+  await promise.then(
+    () => {
+      throw new Error("Expected promise to reject.");
+    },
+    (error) => {
+      expectRecordFields(error, expected);
+    },
+  );
+}
+
+export function mockCallArg(
+  mock: ReturnType<typeof vi.fn>,
+  callIndex = 0,
+): Record<string, unknown> {
+  const call = mock.mock.calls[callIndex];
+  if (!call) {
+    throw new Error(`Expected mock call ${callIndex}`);
+  }
+  return call[0] as Record<string, unknown>;
+}
+
+export function mockCallArgs(mock: ReturnType<typeof vi.fn>): Array<Record<string, unknown>> {
+  return mock.mock.calls.map((call) => call[0] as Record<string, unknown>);
+}
+
+export function findMockCallFields(
+  mock: ReturnType<typeof vi.fn>,
+  expected: Record<string, unknown>,
+) {
+  return mockCallArgs(mock).find((actual) =>
+    Object.entries(expected).every(([key, value]) => Object.is(actual[key], value)),
+  );
+}
+
+export function expectMockCallFields(
+  mock: ReturnType<typeof vi.fn>,
+  expected: Record<string, unknown>,
+) {
+  if (!findMockCallFields(mock, expected)) {
+    throw new Error(`Expected mock call ${JSON.stringify(expected)}`);
+  }
+}
+
+export function expectNoMockCallFields(
+  mock: ReturnType<typeof vi.fn>,
+  expected: Record<string, unknown>,
+) {
+  expect(findMockCallFields(mock, expected)).toBeUndefined();
+}
+
+export function requireTaskByRunId(runId: string) {
+  const task = findTaskByRunId(runId);
+  if (!task) {
+    throw new Error(`Expected task for run ${runId}`);
+  }
+  return task;
+}
+
+export function createRuntime(): {
+  runtime: AcpRuntime;
+  ensureSession: ReturnType<typeof vi.fn>;
+  runTurn: ReturnType<typeof vi.fn>;
+  prepareFreshSession: ReturnType<typeof vi.fn>;
+  cancel: ReturnType<typeof vi.fn>;
+  close: ReturnType<typeof vi.fn>;
+  getCapabilities: ReturnType<typeof vi.fn>;
+  getStatus: ReturnType<typeof vi.fn>;
+  setMode: ReturnType<typeof vi.fn>;
+  setConfigOption: ReturnType<typeof vi.fn>;
+} {
+  const ensureSession = vi.fn(
+    async (input: {
+      sessionKey: string;
+      agent: string;
+      mode: "persistent" | "oneshot";
+      model?: string;
+      thinking?: string;
+      cwd?: string;
+      resumeSessionId?: string;
+    }) => ({
+      sessionKey: input.sessionKey,
+      backend: "acpx",
+      runtimeSessionName: `${input.sessionKey}:${input.mode}:runtime`,
+    }),
+  );
+  const runTurn = vi.fn(async function* () {
+    yield { type: "done" as const };
+  });
+  const prepareFreshSession = vi.fn(async () => {});
+  const cancel = vi.fn(async () => {});
+  const close = vi.fn(async () => {});
+  const getCapabilities = vi.fn(
+    async (): Promise<AcpRuntimeCapabilities> => ({
+      controls: ["session/set_mode", "session/set_config_option", "session/status"],
+    }),
+  );
+  const getStatus = vi.fn(async () => ({
+    summary: "status=alive",
+    details: { status: "alive" },
+  }));
+  const setMode = vi.fn(async () => {});
+  const setConfigOption = vi.fn(async () => {});
+  return {
+    runtime: {
+      ensureSession,
+      runTurn,
+      getCapabilities,
+      getStatus,
+      setMode,
+      setConfigOption,
+      prepareFreshSession,
+      cancel,
+      close,
+    },
+    ensureSession,
+    runTurn,
+    prepareFreshSession,
+    cancel,
+    close,
+    getCapabilities,
+    getStatus,
+    setMode,
+    setConfigOption,
+  };
+}
+
+export function readySessionMeta(overrides: Partial<SessionAcpMeta> = {}): SessionAcpMeta {
+  return {
+    backend: "acpx",
+    agent: "codex",
+    runtimeSessionName: "runtime-1",
+    mode: "persistent" as const,
+    state: "idle" as const,
+    lastActivityAt: Date.now(),
+    ...overrides,
+  };
+}
+
+export function extractStatesFromUpserts(): SessionAcpMeta["state"][] {
+  const states: SessionAcpMeta["state"][] = [];
+  for (const [firstArg] of hoisted.upsertAcpSessionMetaMock.mock.calls) {
+    const payload = firstArg as {
+      mutate: (
+        current: SessionAcpMeta | undefined,
+        entry: { acp?: SessionAcpMeta } | undefined,
+      ) => SessionAcpMeta | null | undefined;
+    };
+    const current = readySessionMeta();
+    const next = payload.mutate(current, { acp: current });
+    if (next?.state) {
+      states.push(next.state);
+    }
+  }
+  return states;
+}
+
+export function extractStateUpsertPersistenceOptions(): Array<{
+  state: SessionAcpMeta["state"];
+  skipMaintenance?: boolean;
+  takeCacheOwnership?: boolean;
+}> {
+  const options: Array<{
+    state: SessionAcpMeta["state"];
+    skipMaintenance?: boolean;
+    takeCacheOwnership?: boolean;
+  }> = [];
+  for (const [firstArg] of hoisted.upsertAcpSessionMetaMock.mock.calls) {
+    const payload = firstArg as {
+      skipMaintenance?: boolean;
+      takeCacheOwnership?: boolean;
+      mutate: (
+        current: SessionAcpMeta | undefined,
+        entry: { acp?: SessionAcpMeta } | undefined,
+      ) => SessionAcpMeta | null | undefined;
+    };
+    const current = readySessionMeta();
+    const next = payload.mutate(current, { acp: current });
+    if (next?.state && payload.skipMaintenance && payload.takeCacheOwnership) {
+      options.push({
+        state: next.state,
+        skipMaintenance: true,
+        takeCacheOwnership: true,
+      });
+    }
+  }
+  return options;
+}
+
+export function extractRuntimeOptionsFromUpserts(): Array<AcpSessionRuntimeOptions | undefined> {
+  const options: Array<AcpSessionRuntimeOptions | undefined> = [];
+  for (const [firstArg] of hoisted.upsertAcpSessionMetaMock.mock.calls) {
+    const payload = firstArg as {
+      mutate: (
+        current: SessionAcpMeta | undefined,
+        entry: { acp?: SessionAcpMeta } | undefined,
+      ) => SessionAcpMeta | null | undefined;
+    };
+    const current = readySessionMeta();
+    const next = payload.mutate(current, { acp: current });
+    if (next) {
+      options.push(next.runtimeOptions);
+    }
+  }
+  return options;
+}
+
+export function installAcpSessionManagerTestLifecycle(): void {
+  beforeEach(() => {
+    resetAcpSessionManagerForTests();
+    resetAcpActiveTurnsForTests();
+    vi.useRealTimers();
+    hoisted.listAcpSessionEntriesMock.mockReset().mockResolvedValue([]);
+    hoisted.readAcpSessionEntryMock.mockReset();
+    hoisted.upsertAcpSessionMetaMock.mockReset().mockResolvedValue(null);
+    hoisted.requireAcpRuntimeBackendMock.mockReset();
+    hoisted.getAcpRuntimeBackendMock.mockReset().mockImplementation((backendId?: string) => {
+      try {
+        return hoisted.requireAcpRuntimeBackendMock(backendId);
+      } catch {
+        return null;
+      }
+    });
+  });
+
+  afterEach(() => {
+    if (ORIGINAL_STATE_DIR === undefined) {
+      delete process.env.OPENCLAW_STATE_DIR;
+    } else {
+      process.env.OPENCLAW_STATE_DIR = ORIGINAL_STATE_DIR;
+    }
+    resetHeartbeatWakeStateForTests();
+    resetTaskRegistryForTests({ persist: false });
+    resetTaskFlowRegistryForTests({ persist: false });
+  });
+}
