+++ src/acp/control-plane/manager.test.ts
@@ -1,338 +1,38 @@
 import { setTimeout as scheduleNativeTimeout } from "node:timers";
 import { setTimeout as sleep } from "node:timers/promises";
-import type { AcpRuntime, AcpRuntimeCapabilities } from "@openclaw/acp-core/runtime/types";
 import { MAX_TIMER_TIMEOUT_MS } from "@openclaw/normalization-core/number-coercion";
-import { afterEach, beforeEach, describe, expect, it, vi } from "vitest";
-import type { OpenClawConfig } from "../../config/config.js";
-import type { AcpSessionRuntimeOptions, SessionAcpMeta } from "../../config/sessions/types.js";
-import { resetHeartbeatWakeStateForTests } from "../../infra/heartbeat-wake.js";
-import { withTempDir } from "../../test-helpers/temp-dir.js";
-import { isAcpTurnActive, resetAcpActiveTurnsForTests } from "./active-turns.js";
-
-const hoisted = vi.hoisted(() => {
-  const listAcpSessionEntriesMock = vi.fn();
-  const readAcpSessionEntryMock = vi.fn();
-  const upsertAcpSessionMetaMock = vi.fn();
-  const getAcpRuntimeBackendMock = vi.fn();
-  const requireAcpRuntimeBackendMock = vi.fn();
-  return {
-    listAcpSessionEntriesMock,
-    readAcpSessionEntryMock,
-    upsertAcpSessionMetaMock,
-    getAcpRuntimeBackendMock,
-    requireAcpRuntimeBackendMock,
-  };
-});
-
-vi.mock("../runtime/session-meta.js", () => ({
-  listAcpSessionEntries: (params: unknown) => hoisted.listAcpSessionEntriesMock(params),
-  readAcpSessionEntry: (params: unknown) => hoisted.readAcpSessionEntryMock(params),
-  upsertAcpSessionMeta: (params: unknown) => hoisted.upsertAcpSessionMetaMock(params),
-}));
-
-vi.mock("../runtime/registry.js", () => ({
-  getAcpRuntimeBackend: (backendId?: string) => hoisted.getAcpRuntimeBackendMock(backendId),
-  requireAcpRuntimeBackend: (backendId?: string) => hoisted.requireAcpRuntimeBackendMock(backendId),
-}));
-
-const {
+import { describe, expect, it, vi } from "vitest";
+import { isAcpTurnActive } from "./active-turns.js";
+import {
+  type AcpRuntime,
+  AcpRuntimeError,
   AcpSessionManager,
-  testing: { resetAcpSessionManagerForTests },
-} = await import("./manager.js");
-const { AcpRuntimeError } = await import("../runtime/errors.js");
-const { findTaskByRunId, resetTaskRegistryForTests } = await import("../../tasks/task-registry.js");
-const { resetTaskFlowRegistryForTests } = await import("../../tasks/task-flow-registry.js");
-const { configureTaskFlowRegistryRuntime } =
-  await import("../../tasks/task-flow-registry.store.js");
-const { installInMemoryTaskRegistryRuntime } =
-  await import("../../test-utils/task-registry-runtime.js");
-
-const baseCfg = {
-  acp: {
-    enabled: true,
-    backend: "acpx",
-    dispatch: { enabled: true },
-  },
-} as const;
-const ORIGINAL_STATE_DIR = process.env.OPENCLAW_STATE_DIR;
-
-async function withAcpManagerTaskStateDir(run: (root: string) => Promise<void>): Promise<void> {
-  await withTempDir({ prefix: "openclaw-acp-manager-task-" }, async (root) => {
-    process.env.OPENCLAW_STATE_DIR = root;
-    resetTaskRegistryForTests({ persist: false });
-    resetTaskFlowRegistryForTests({ persist: false });
-    installInMemoryTaskRegistryRuntime();
-    configureTaskFlowRegistryRuntime({
-      store: {
-        loadSnapshot: () => ({
-          flows: new Map(),
-        }),
-        saveSnapshot: () => {},
-        upsertFlow: () => {},
-        deleteFlow: () => {},
-        close: () => {},
-      },
-    });
-    try {
-      await run(root);
-    } finally {
-      resetTaskRegistryForTests({ persist: false });
-      resetTaskFlowRegistryForTests({ persist: false });
-    }
-  });
-}
-
-async function flushMicrotasks(rounds = 3): Promise<void> {
-  for (let index = 0; index < rounds; index += 1) {
-    await Promise.resolve();
-  }
-}
-
-function createDeferred(): { promise: Promise<void>; resolve: () => void } {
-  let resolve: (() => void) | undefined;
-  const promise = new Promise<void>((next) => {
-    resolve = next;
-  });
-  if (!resolve) {
-    throw new Error("Expected deferred resolver to be initialized");
-  }
-  return { promise, resolve };
-}
-
-function expectRecordFields(record: unknown, expected: Record<string, unknown>) {
-  if (!record || typeof record !== "object") {
-    throw new Error("Expected record");
-  }
-  const actual = record as Record<string, unknown>;
-  for (const [key, value] of Object.entries(expected)) {
-    expect(actual[key]).toEqual(value);
-  }
-  return actual;
-}
-
-async function expectRejectedRecord(promise: Promise<unknown>, expected: Record<string, unknown>) {
-  await promise.then(
-    () => {
-      throw new Error("Expected promise to reject.");
-    },
-    (error) => {
-      expectRecordFields(error, expected);
-    },
-  );
-}
-
-function mockCallArg(mock: ReturnType<typeof vi.fn>, callIndex = 0): Record<string, unknown> {
-  const call = mock.mock.calls[callIndex];
-  if (!call) {
-    throw new Error(`Expected mock call ${callIndex}`);
-  }
-  return call[0] as Record<string, unknown>;
-}
-
-function mockCallArgs(mock: ReturnType<typeof vi.fn>): Array<Record<string, unknown>> {
-  return mock.mock.calls.map((call) => call[0] as Record<string, unknown>);
-}
-
-function findMockCallFields(mock: ReturnType<typeof vi.fn>, expected: Record<string, unknown>) {
-  return mockCallArgs(mock).find((actual) =>
-    Object.entries(expected).every(([key, value]) => Object.is(actual[key], value)),
-  );
-}
-
-function expectMockCallFields(mock: ReturnType<typeof vi.fn>, expected: Record<string, unknown>) {
-  if (!findMockCallFields(mock, expected)) {
-    throw new Error(`Expected mock call ${JSON.stringify(expected)}`);
-  }
-}
-
-function expectNoMockCallFields(mock: ReturnType<typeof vi.fn>, expected: Record<string, unknown>) {
-  expect(findMockCallFields(mock, expected)).toBeUndefined();
-}
-
-function requireTaskByRunId(runId: string) {
-  const task = findTaskByRunId(runId);
-  if (!task) {
-    throw new Error(`Expected task for run ${runId}`);
-  }
-  return task;
-}
-
-function createRuntime(): {
-  runtime: AcpRuntime;
-  ensureSession: ReturnType<typeof vi.fn>;
-  runTurn: ReturnType<typeof vi.fn>;
-  prepareFreshSession: ReturnType<typeof vi.fn>;
-  cancel: ReturnType<typeof vi.fn>;
-  close: ReturnType<typeof vi.fn>;
-  getCapabilities: ReturnType<typeof vi.fn>;
-  getStatus: ReturnType<typeof vi.fn>;
-  setMode: ReturnType<typeof vi.fn>;
-  setConfigOption: ReturnType<typeof vi.fn>;
-} {
-  const ensureSession = vi.fn(
-    async (input: {
-      sessionKey: string;
-      agent: string;
-      mode: "persistent" | "oneshot";
-      model?: string;
-      thinking?: string;
-      cwd?: string;
-      resumeSessionId?: string;
-    }) => ({
-      sessionKey: input.sessionKey,
-      backend: "acpx",
-      runtimeSessionName: `${input.sessionKey}:${input.mode}:runtime`,
-    }),
-  );
-  const runTurn = vi.fn(async function* () {
-    yield { type: "done" as const };
-  });
-  const prepareFreshSession = vi.fn(async () => {});
-  const cancel = vi.fn(async () => {});
-  const close = vi.fn(async () => {});
-  const getCapabilities = vi.fn(
-    async (): Promise<AcpRuntimeCapabilities> => ({
-      controls: ["session/set_mode", "session/set_config_option", "session/status"],
-    }),
-  );
-  const getStatus = vi.fn(async () => ({
-    summary: "status=alive",
-    details: { status: "alive" },
-  }));
-  const setMode = vi.fn(async () => {});
-  const setConfigOption = vi.fn(async () => {});
-  return {
-    runtime: {
-      ensureSession,
-      runTurn,
-      getCapabilities,
-      getStatus,
-      setMode,
-      setConfigOption,
-      prepareFreshSession,
-      cancel,
-      close,
-    },
-    ensureSession,
-    runTurn,
-    prepareFreshSession,
-    cancel,
-    close,
-    getCapabilities,
-    getStatus,
-    setMode,
-    setConfigOption,
-  };
-}
-
-function readySessionMeta(overrides: Partial<SessionAcpMeta> = {}): SessionAcpMeta {
-  return {
-    backend: "acpx",
-    agent: "codex",
-    runtimeSessionName: "runtime-1",
-    mode: "persistent" as const,
-    state: "idle" as const,
-    lastActivityAt: Date.now(),
-    ...overrides,
-  };
-}
-
-function extractStatesFromUpserts(): SessionAcpMeta["state"][] {
-  const states: SessionAcpMeta["state"][] = [];
-  for (const [firstArg] of hoisted.upsertAcpSessionMetaMock.mock.calls) {
-    const payload = firstArg as {
-      mutate: (
-        current: SessionAcpMeta | undefined,
-        entry: { acp?: SessionAcpMeta } | undefined,
-      ) => SessionAcpMeta | null | undefined;
-    };
-    const current = readySessionMeta();
-    const next = payload.mutate(current, { acp: current });
-    if (next?.state) {
-      states.push(next.state);
-    }
-  }
-  return states;
-}
-
-function extractStateUpsertPersistenceOptions(): Array<{
-  state: SessionAcpMeta["state"];
-  skipMaintenance?: boolean;
-  takeCacheOwnership?: boolean;
-}> {
-  const options: Array<{
-    state: SessionAcpMeta["state"];
-    skipMaintenance?: boolean;
-    takeCacheOwnership?: boolean;
-  }> = [];
-  for (const [firstArg] of hoisted.upsertAcpSessionMetaMock.mock.calls) {
-    const payload = firstArg as {
-      skipMaintenance?: boolean;
-      takeCacheOwnership?: boolean;
-      mutate: (
-        current: SessionAcpMeta | undefined,
-        entry: { acp?: SessionAcpMeta } | undefined,
-      ) => SessionAcpMeta | null | undefined;
-    };
-    const current = readySessionMeta();
-    const next = payload.mutate(current, { acp: current });
-    if (next?.state && payload.skipMaintenance && payload.takeCacheOwnership) {
-      options.push({
-        state: next.state,
-        skipMaintenance: true,
-        takeCacheOwnership: true,
-      });
-    }
-  }
-  return options;
-}
-
-function extractRuntimeOptionsFromUpserts(): Array<AcpSessionRuntimeOptions | undefined> {
-  const options: Array<AcpSessionRuntimeOptions | undefined> = [];
-  for (const [firstArg] of hoisted.upsertAcpSessionMetaMock.mock.calls) {
-    const payload = firstArg as {
-      mutate: (
-        current: SessionAcpMeta | undefined,
-        entry: { acp?: SessionAcpMeta } | undefined,
-      ) => SessionAcpMeta | null | undefined;
-    };
-    const current = readySessionMeta();
-    const next = payload.mutate(current, { acp: current });
-    if (next) {
-      options.push(next.runtimeOptions);
-    }
-  }
-  return options;
-}
+  baseCfg,
+  createDeferred,
+  createRuntime,
+  expectMockCallFields,
+  expectNoMockCallFields,
+  expectRecordFields,
+  expectRejectedRecord,
+  extractRuntimeOptionsFromUpserts,
+  extractStateUpsertPersistenceOptions,
+  extractStatesFromUpserts,
+  findTaskByRunId,
+  flushMicrotasks,
+  hoisted,
+  installAcpSessionManagerTestLifecycle,
+  mockCallArg,
+  mockCallArgs,
+  readySessionMeta,
+  requireTaskByRunId,
+  type OpenClawConfig,
+  resetAcpSessionManagerForTests,
+  type SessionAcpMeta,
+  withAcpManagerTaskStateDir,
+} from "./manager.test-helpers.js";
 
 describe("AcpSessionManager", () => {
-  beforeEach(() => {
-    resetAcpSessionManagerForTests();
-    resetAcpActiveTurnsForTests();
-    vi.useRealTimers();
-    hoisted.listAcpSessionEntriesMock.mockReset().mockResolvedValue([]);
-    hoisted.readAcpSessionEntryMock.mockReset();
-    hoisted.upsertAcpSessionMetaMock.mockReset().mockResolvedValue(null);
-    hoisted.requireAcpRuntimeBackendMock.mockReset();
-    hoisted.getAcpRuntimeBackendMock.mockReset().mockImplementation((backendId?: string) => {
-      try {
-        return hoisted.requireAcpRuntimeBackendMock(backendId);
-      } catch {
-        return null;
-      }
-    });
-  });
-
-  afterEach(() => {
-    if (ORIGINAL_STATE_DIR === undefined) {
-      delete process.env.OPENCLAW_STATE_DIR;
-    } else {
-      process.env.OPENCLAW_STATE_DIR = ORIGINAL_STATE_DIR;
-    }
-    resetHeartbeatWakeStateForTests();
-    resetTaskRegistryForTests({ persist: false });
-    resetTaskFlowRegistryForTests({ persist: false });
-  });
+  installAcpSessionManagerTestLifecycle();
 
   it("marks ACP-shaped sessions without metadata as stale", () => {
     hoisted.readAcpSessionEntryMock.mockReturnValue(null);
@@ -1898,249 +1598,6 @@ describe("AcpSessionManager", () => {
 
     expect(hoisted.requireAcpRuntimeBackendMock).toHaveBeenCalledWith("metadata-backend");
     expect(runtimeState.runTurn).toHaveBeenCalledTimes(1);
-  });
-
-  function setupFailoverBackends(
-    params: {
-      initialBackend?: "primary-backend" | "fallback-backend";
-      primaryUnavailableError?: Error;
-    } = {},
-  ) {
-    const primaryRuntime = createRuntime();
-    const fallbackRuntime = createRuntime();
-    const sessionKey = "agent:codex:acp:session-1";
-    const initialBackend = params.initialBackend ?? "primary-backend";
-    let currentMeta = readySessionMeta({
-      backend: initialBackend,
-      runtimeSessionName:
-        initialBackend === "fallback-backend" ? "fallback-runtime" : "primary-runtime",
-    });
-    primaryRuntime.ensureSession.mockImplementation(async (input) => ({
-      sessionKey: input.sessionKey,
-      backend: "primary-backend",
-      runtimeSessionName: "primary-runtime",
-    }));
-    fallbackRuntime.ensureSession.mockImplementation(async (input) => ({
-      sessionKey: input.sessionKey,
-      backend: "fallback-backend",
-      runtimeSessionName: "fallback-runtime",
-    }));
-    hoisted.requireAcpRuntimeBackendMock.mockImplementation((backendId?: string) => {
-      if (backendId === "primary-backend") {
-        if (params.primaryUnavailableError) {
-          throw params.primaryUnavailableError;
-        }
-        return {
-          id: "primary-backend",
-          runtime: primaryRuntime.runtime,
-        };
-      }
-      if (backendId === "fallback-backend") {
-        return {
-          id: "fallback-backend",
-          runtime: fallbackRuntime.runtime,
-        };
-      }
-      throw new Error(`unexpected backend ${backendId ?? "<auto>"}`);
-    });
-    hoisted.readAcpSessionEntryMock.mockImplementation(() => ({
-      sessionKey,
-      storeSessionKey: sessionKey,
-      acp: currentMeta,
-    }));
-    hoisted.upsertAcpSessionMetaMock.mockImplementation(async (paramsUnknown: unknown) => {
-      const upsertParams = paramsUnknown as {
-        mutate: (
-          current: SessionAcpMeta | undefined,
-          entry: { acp?: SessionAcpMeta } | undefined,
-        ) => SessionAcpMeta | null | undefined;
-      };
-      const next = upsertParams.mutate(currentMeta, { acp: currentMeta });
-      if (next) {
-        currentMeta = next;
-      }
-      return {
-        sessionId: "session-1",
-        updatedAt: Date.now(),
-        acp: currentMeta,
-      };
-    });
-    const cfg = {
-      acp: {
-        ...baseCfg.acp,
-        backend: "primary-backend",
-        fallbacks: ["fallback-backend"],
-      },
-    } as OpenClawConfig;
-    return {
-      cfg,
-      fallbackRuntime,
-      get currentMeta() {
-        return currentMeta;
-      },
-      primaryRuntime,
-      sessionKey,
-    };
-  }
-
-  it("starts later failover turns on the configured primary backend", async () => {
-    const harness = setupFailoverBackends({ initialBackend: "fallback-backend" });
-
-    const manager = new AcpSessionManager();
-    await manager.runTurn({
-      cfg: harness.cfg,
-      sessionKey: harness.sessionKey,
-      text: "use primary",
-      mode: "prompt",
-      requestId: "r-primary",
-    });
-
-    expect(hoisted.requireAcpRuntimeBackendMock).toHaveBeenCalledWith("primary-backend");
-    expect(harness.primaryRuntime.runTurn).toHaveBeenCalledTimes(1);
-    expect(harness.fallbackRuntime.runTurn).not.toHaveBeenCalled();
-    expect(harness.currentMeta.backend).toBe("primary-backend");
-  });
-
-  it("closes cached fallback handles before returning later turns to the primary backend", async () => {
-    const harness = setupFailoverBackends();
-    harness.primaryRuntime.runTurn.mockImplementationOnce(async function* () {
-      if (Date.now() < 0) {
-        yield { type: "done" as const };
-      }
-      throw new AcpRuntimeError("ACP_TURN_FAILED", "backend unavailable");
-    });
-
-    const manager = new AcpSessionManager();
-    await manager.runTurn({
-      cfg: harness.cfg,
-      sessionKey: harness.sessionKey,
-      text: "use fallback",
-      mode: "prompt",
-      requestId: "r-fallback",
-    });
-    expect(harness.currentMeta.backend).toBe("fallback-backend");
-    expect(harness.fallbackRuntime.runTurn).toHaveBeenCalledTimes(1);
-
-    harness.fallbackRuntime.close.mockClear();
-    await manager.runTurn({
-      cfg: harness.cfg,
-      sessionKey: harness.sessionKey,
-      text: "return to primary",
-      mode: "prompt",
-      requestId: "r-primary",
-    });
-
-    expect(harness.fallbackRuntime.close).toHaveBeenCalledWith(
-      expect.objectContaining({
-        reason: "runtime-handle-replaced",
-      }),
-    );
-    expect(harness.primaryRuntime.runTurn).toHaveBeenCalledTimes(2);
-    expect(harness.currentMeta.backend).toBe("primary-backend");
-  });
-
-  it("closes the previous persistent handle before switching fallback backends", async () => {
-    const harness = setupFailoverBackends();
-    harness.primaryRuntime.runTurn.mockImplementation(async function* () {
-      if (Date.now() < 0) {
-        yield { type: "done" as const };
-      }
-      throw new AcpRuntimeError("ACP_TURN_FAILED", "backend unavailable");
-    });
-
-    const manager = new AcpSessionManager();
-    await expect(
-      manager.runTurn({
-        cfg: harness.cfg,
-        sessionKey: harness.sessionKey,
-        text: "fallback",
-        mode: "prompt",
-        requestId: "r-fallback",
-      }),
-    ).resolves.toBeUndefined();
-
-    expect(harness.primaryRuntime.close).toHaveBeenCalledWith(
-      expect.objectContaining({
-        reason: "backend-failover",
-      }),
-    );
-    expect(harness.fallbackRuntime.runTurn).toHaveBeenCalledTimes(1);
-  });
-
-  it("fails over when the primary backend is registered but unavailable", async () => {
-    const harness = setupFailoverBackends({
-      primaryUnavailableError: new AcpRuntimeError(
-        "ACP_BACKEND_UNAVAILABLE",
-        "primary backend unavailable",
-      ),
-    });
-
-    const manager = new AcpSessionManager();
-    await expect(
-      manager.runTurn({
-        cfg: harness.cfg,
-        sessionKey: harness.sessionKey,
-        text: "fallback",
-        mode: "prompt",
-        requestId: "r-unavailable",
-      }),
-    ).resolves.toBeUndefined();
-
-    expect(hoisted.requireAcpRuntimeBackendMock).toHaveBeenCalledWith("primary-backend");
-    expect(hoisted.requireAcpRuntimeBackendMock).toHaveBeenCalledWith("fallback-backend");
-    expect(harness.fallbackRuntime.runTurn).toHaveBeenCalledTimes(1);
-  });
-
-  it("fails over for common rate limit wording before output", async () => {
-    const harness = setupFailoverBackends();
-    harness.primaryRuntime.runTurn.mockImplementation(async function* () {
-      if (Date.now() < 0) {
-        yield { type: "done" as const };
-      }
-      throw new AcpRuntimeError("ACP_TURN_FAILED", "rate limit exceeded");
-    });
-
-    const manager = new AcpSessionManager();
-    await expect(
-      manager.runTurn({
-        cfg: harness.cfg,
-        sessionKey: harness.sessionKey,
-        text: "fallback",
-        mode: "prompt",
-        requestId: "r-rate-limit",
-      }),
-    ).resolves.toBeUndefined();
-
-    expect(harness.primaryRuntime.runTurn).toHaveBeenCalledTimes(1);
-    expect(harness.fallbackRuntime.runTurn).toHaveBeenCalledTimes(1);
-  });
-
-  it("does not fail over after a backend has emitted output", async () => {
-    const harness = setupFailoverBackends();
-    harness.primaryRuntime.runTurn.mockImplementation(async function* () {
-      yield { type: "text_delta" as const, text: "partial" };
-      throw new AcpRuntimeError("ACP_TURN_FAILED", "backend unavailable");
-    });
-    const events: unknown[] = [];
-
-    const manager = new AcpSessionManager();
-    await expect(
-      manager.runTurn({
-        cfg: harness.cfg,
-        sessionKey: harness.sessionKey,
-        text: "do not duplicate",
-        mode: "prompt",
-        requestId: "r-output",
-        onEvent: (event) => {
-          events.push(event);
-        },
-      }),
-    ).rejects.toMatchObject({
-      code: "ACP_TURN_FAILED",
-    });
-
-    expect(events).toEqual([expect.objectContaining({ type: "text_delta", text: "partial" })]);
-    expect(harness.fallbackRuntime.runTurn).not.toHaveBeenCalled();
   });
 
   it("persists runtime options provided during initializeSession", async () => {
