+++ test/vitest/vitest.unit-fast-paths.mjs
@@ -69,6 +69,7 @@ export const forcedUnitFastTestFiles = [
   "packages/memory-host-sdk/src/host/session-files.test.ts",
   "src/acp/client.test.ts",
   "src/acp/control-plane/manager.backend-failover.test.ts",
+  "src/acp/control-plane/manager.failover.test.ts",
   "src/acp/control-plane/manager.test.ts",
   "src/acp/session-mapper.test.ts",
   "src/acp/persistent-bindings.lifecycle.test.ts",
